chore: refactor account preferences to typescript
